+++ MyOwnOnlineShop/WebAPI/Functions/Commands/ProductCommnds/DeleteProductHandler.cs
@@ -1,5 +1,6 @@
 ﻿using Application.Interfaces;
 using MediatR;
+using WebAPI.Wrappers;
 
 namespace WebAPI.Functions.Commands.ProductCommnds;
 
@@ -7,6 +8,12 @@ public class DeleteProductHandler(IProductService productService) : IRequestHand
 {
     public async Task<Unit> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
     {
+        var product = await productService.GetProductById(request.Id);
+        if (product == null)
+        {
+            throw new NotFoundException($"Product with id {request.Id} does not exist.");
+        }
+
         await productService.DeleteProduct(request.Id);
         return Unit.Value;
     }
